requirements.txt
pymodbus>=3.0.0
pyserial>=3.5
pyyaml>=6.0
flask>=3.0.0
